+++ src/main/java/Dusigi/bookworm/domain/bookshelf/data/dto/request/BookDetailDto.java
@@ -1,20 +1,18 @@
-package Dusigi.bookworm.domain.bookshelf.data.dto;
+package Dusigi.bookworm.domain.bookshelf.data.dto.request;
 
-import io.lettuce.core.CompositeArgument;
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
-public class BookDto {
+@NoArgsConstructor
+public class BookDetailDto {
     @NotBlank(message = "저자는 공백을 허용하지 않습니다.")
-    String author;
+    private String author;
     @NotBlank(message = "제목은 공백을 허용하지 않습니다.")
-    String title;
+    private String title;
     @NotBlank(message = "내용은 공백을 허용하지 않습니다.")
-    String text;
-    @NotBlank(message = "책장은 공백을 허용하지 않습니다.")
-    String bookshelf;
+    private String text;
 }
