+++ src/main/java/Dusigi/bookworm/domain/bookshelf/data/dto/request/BookDto.java
@@ -1,18 +1,20 @@
-package Dusigi.bookworm.domain.bookshelf.data.dto;
+package Dusigi.bookworm.domain.bookshelf.data.dto.request;
 
+import io.lettuce.core.CompositeArgument;
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
-@NoArgsConstructor
-public class BookDetailDto {
+public class BookDto {
     @NotBlank(message = "저자는 공백을 허용하지 않습니다.")
-    private String author;
+    String author;
     @NotBlank(message = "제목은 공백을 허용하지 않습니다.")
-    private String title;
+    String title;
     @NotBlank(message = "내용은 공백을 허용하지 않습니다.")
-    private String text;
+    String text;
+    @NotBlank(message = "책장은 공백을 허용하지 않습니다.")
+    String bookshelf;
 }
