+++ src/main/java/Dusigi/bookworm/domain/bookshelf/service/BookDetailService.java
@@ -1,6 +1,6 @@
 package Dusigi.bookworm.domain.bookshelf.service;
 
-import Dusigi.bookworm.domain.bookshelf.data.dto.BookDetailDto;
+import Dusigi.bookworm.domain.bookshelf.data.dto.request.BookDetailDto;
 
 public interface BookDetailService {
     BookDetailDto getBook(Long id);
