+++ src/main/java/Dusigi/bookworm/domain/bookshelf/service/BookService.java
@@ -1,7 +1,6 @@
 package Dusigi.bookworm.domain.bookshelf.service;
 
-import Dusigi.bookworm.domain.bookshelf.data.dto.BookDetailDto;
-import Dusigi.bookworm.domain.bookshelf.data.dto.BookDto;
+import Dusigi.bookworm.domain.bookshelf.data.dto.request.BookDto;
 import Dusigi.bookworm.domain.bookshelf.data.entity.Book;
