+++ src/main/java/Dusigi/bookworm/domain/bookshelf/service/Impl/BookServiceImpl.java
@@ -2,7 +2,7 @@ package Dusigi.bookworm.domain.bookshelf.service.Impl;
 
 import Dusigi.bookworm.domain.bookshelf.data.repository.BookRepository;
 import Dusigi.bookworm.domain.bookshelf.service.BookService;
-import Dusigi.bookworm.domain.bookshelf.data.dto.BookDto;
+import Dusigi.bookworm.domain.bookshelf.data.dto.request.BookDto;
 import Dusigi.bookworm.domain.bookshelf.data.entity.Book;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
