+++ src/main/java/Dusigi/bookworm/domain/bookshelf/service/Impl/BookshelfServiceImpl.java
@@ -1,6 +1,6 @@
 package Dusigi.bookworm.domain.bookshelf.service.Impl;
 
-import Dusigi.bookworm.domain.bookshelf.data.dto.BookshelfDto;
+import Dusigi.bookworm.domain.bookshelf.data.dto.response.BookshelfDto;
 import Dusigi.bookworm.domain.bookshelf.data.entity.Book;
 import Dusigi.bookworm.domain.bookshelf.data.repository.BookRepository;
 import Dusigi.bookworm.domain.bookshelf.service.BookshelfService;
